requirements.txt
lxml==4.7.1
mysql-connector-python==8.0.27
paramiko==2.9.1
requests==2.26.0
safetywrap==1.5.0
tqdm==4.62.3
